+++ HISTORY.rst
@@ -11,6 +11,7 @@ Unreleased
 * Change admin password at the end of setup
 * Update all modules (odoo and oca) at last version
 * Card 99: Install module to activate authentication by google account
+* Card 216: On payment, set partner and ref even on VAT item entries
 
 **Bugfixes**